+++ apps/examples/dpp/main.c
@@ -100,7 +100,7 @@ AM_NORETURN void am_assert_failure(
         line,
         am_pal_task_get_own_id()
     );
-    am_event_log_pools_unsafe(/*num=*/-1, log_pool);
+    am_event_pool_log_unsafe(/*num=*/-1, log_pool);
     am_ao_log_event_queues_unsafe(/*num=*/-1, log_queue);
     am_pal_flush();
     __builtin_trap();