+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Rename `am_event_get_pool_nfree()` to `am_event_pool_get_nfree()`.
 - Rename `am_event_get_pool_nblocks()` to `am_event_pool_get_nblocks()`.
 - Rename `am_event_get_npools()` to `am_event_pool_get_num()`.
+- Rename `am_event_log_pools_unsafe()` to `am_event_pool_log_unsafe()`.
 
 ### Added